[REF] stock: scrap in receipt
In the receipts, you can only scrap after the reception. This implies
that the source location of the scrap move is the destination location
of the picking.

Create the scrap move as draft, else in the receipt it used the state of
the picking (done) and calling _action_done again resulted in double
valuation layers

task-2042399

closes odoo/odoo#35179

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>