[FIX] website: only stop clone's animations on cloning
/!\ Should not be forward-ported ! As further explained, this commit is
/!\ only partial backports of 12.0 commits. All conflicts should be
/!\ ignored.

A recently merged PR https://github.com/odoo/odoo/pull/35029 was in
fact developed for 12.0 originally and finally merged in 11.0 to solve
potential same bugs. The problem is that one line of that merge relied
on a 12.0 feature that is missing in 11.0: the ability to stop
animations *on part of the DOM only*.

This commit chose to backport that 12.0 feature instead of retargeting
the original PR as it will be needed for another unrelated fixes. So,
this commit partially backport those commits:
- https://github.com/odoo/odoo/commit/4967e22bd10a9494d290aafcff61b7acb6ec40f7
- https://github.com/odoo/odoo/commit/3fd206b19cc5e8aa0c8885da3f65e280c941eb22

(Only the stop animation ability part, not the way it was used).

Note that the first commit was already partially backported with
https://github.com/odoo/odoo/commit/be57bc64afd227003435273a6bbf166843165ddd

closes odoo/odoo#35230

Signed-off-by: Romain Derie (rde) <rde@odoo.com>