[FIX] website: restore folded ul option
The option which allows to create foldable lists was not working
anymore. Taking advantage of https://github.com/odoo/odoo/commit/09181ef6b20c36c95d7568deb2157946684d3a4b
solves the issue.

Note: the option should be improved a lot to be more beautiful / useful
but this will be a master task.

closes odoo/odoo#35249

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>