[REF] mail: remove notified_partner_ids from mark_all_as_read
mark_all_as_read is using notified_partner_ids to filter
the messages to use. But anyway, the notification will
be filtered on partner right after that.

Since if we have a domain, the message set will be quite small,
(scope of a thread) we search all those messages.

If not, we actually avoid to search on message and go right
to the notifications.

Reading the message_ids will garantee to have the minimal
set of message that where actually passed from unread to read.
Since data should not be in cache at this time and mark_all_as_read
is usually called alone, we prefer to use cache and perform a read
immediatly.

Task-ID 402597