[IMP] web: show avatar correctly even when rectangle image is used
Before this commit,
In top-right corner, the avatar that is being shown doesn't correctly
look like a circle if the shape of the avatar is rectangle.

After this commit,
The avatar appears in the top-right corner will be shown in correct circle
shape, irrelevant of the image is square or not.

task-2042276
closes: #35265

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>