[FIX] website_sale_wishlist: improve product page loading time
When the product has a lot of variants which are displayed as checkboxes
or as colors, the jQuery selector was matching as many items, and the
event was triggered for each of them.
The result was the change event of website_sale being triggered many
times, where only one call would be enough.

The change replicates what was made on v12.0 on the same line by commit
6ca3c7e674da4f82086207a9d8d9884aaa0db008.

opw 2036356

closes odoo/odoo#35322

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>