[REF] stock_account: _compute_average_price unit price rounding
Since the introduction of the stock valuation layer, unit_cost is
rounded in the company currency (actually it is NOT because of an ORM
bug that will be fixed later). With the test introduced at [0], fixing
that ORM bug shows we have a rounding issue (27.02 instead of 27) due to
the company currency rounding on the unit_cost. Since the logic of the
anglo saxon will probably always give some rounding issue (we always
return a price unit that will be multiply by the invoiced quantity)
recomputing a not rounded price unit by using the not rounded value with
the not rounded quantity seems a good enough compromise.

[0] 3e2f17633787b592061a5c72ec504a3d0ee0afc0

closes odoo/odoo#35384

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>